feat: align SUI input amount
